+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+- Fixed background color of AthleticsEventsPanel [#5871](https://github.com/rokwire/illinois-app/issues/5871).
+- Show bottom tab bar in all new Athletics panels [#5871](https://github.com/rokwire/illinois-app/issues/5871).
 
 ## [8.2.15] - 2026-09-01
 ### Removed